Change bigint to int to save space

+++ SQL Files/createTables.sql
@@ -1,7 +1,7 @@
 create table location
 (
-	locID bigint(20) primary key not null auto_increment,
-	addressID bigint(20) not null,
+	locID int primary key not null auto_increment,
+	addressID int not null,
 	roadChar enum("Not at a Junction", "Driveway", "Parking Lot", "Five-point or more", "T - Intersection", "Four-way Intersection", "On Ramp", "Traffic circle / roundabout",
 	"Y - Intersection", "Off Ramp", "Not Reported"),
 	foreign key(addressID) references address(addressID)
@@ -9,15 +9,15 @@ create table location
 
 create table address
 (
-	addressID bigint(20) primary key not null auto_increment,
+	addressID int primary key not null auto_increment,
 	streetAddress varchar(50) not null,
-	cityID bigint(20) not null,
+	cityID int not null,
 	foreign key(cityID) references city(cityID)
 );
 
 create table weather
 (
-	condID bigint(20) primary key not null auto_increment,
+	condID int primary key not null auto_increment,
 	weather enum("Freezing Precipitation", "Cloudy", "Wind", "Clear", "Rain", "Unknown"),
 	surfaceCond enum("Snow", "Ice", "Wet", "Dry", "Slush", "Sand, mud, dirt, oil, gravel", "Water (standing / moving)", "Unknown"),
 	dayNight enum("Day", "Night", "Unknown")
@@ -26,17 +26,17 @@ create table weather
 
 create table driver
 (
-	driverID bigint(20) primary key not null auto_increment,
+	driverID int primary key not null auto_increment,
 	driverImpair enum("Alcohol", "Drugs", "None"),
 	driverDamage enum("Property Damage Only", "Injury", "Fatal", "Unknown")
 );
 
 create table accident
 (
-	accID bigint(20) primary key not null auto_increment,
-	locID bigint(20) not null,
-	condID bigint(20) not null,
-	driverID bigint(20) not null,
+	accID int primary key not null auto_increment,
+	locID int not null,
+	condID int not null,
+	driverID int not null,
 	foreign key(locID) references location(locID),
 	foreign key(condID) references weather(condID),
 	foreign key(driverID) references driver(driverID),
@@ -52,7 +52,7 @@ create table accident
 
 create table animal
 (
-	accID bigint(20) not null,
+	accID int not null,
 	animalType enum("Deer", "Moose", "Wild", "Domestic") not null,
 	foreign key(accID) references accident(accID),
 	primary key(accID, animalType)
@@ -61,7 +61,7 @@ create table animal
 
 create table vehicle
 (
-	accID bigint(20) not null,
+	accID int not null,
 	vehicleType enum("Heavy Truck", "Bicycle", "Pedestrian", "Motorcycle") not null,
 	foreign key(accID) references accident(accID),
 	primary key(accID, vehicleType)
@@ -70,6 +70,6 @@ create table vehicle
 
 create table city
 (
-	cityID bigint(20) primary key not null auto_increment,
+	cityID int primary key not null auto_increment,
 	cityName varchar(20) not null
 );
